feat: Добавлена обработка доступности погоды для оборудования

+++ src/services/equipment.service.ts
@@ -1,9 +1,13 @@
 import { Equipment } from "../domains/entities/equipment.entity.js";
 import { CreateEquipmentRequest } from "../dto/contracts/create-equipment.request.js";
+import { GetEquipmentWeatherResponse } from "../dto/contracts/get-equipment-weather.response.js";
 import { getEquipmentsRequest } from "../dto/contracts/get-equipments.request.js";
 import { UpdateEquipmentRequest } from "../dto/contracts/update-equipment.request.js";
+import { WeatherRule } from "../dto/weather-rule.js";
 import { IEquipmentRepository } from "../repositories/abstractions/equipment-repository.interface.js";
 import { EquipmentRepositoryMemory } from "../repositories/implementations/in-memory-equipment-repository.js";
+import { weatherSuitableSchema } from "../validators/schemas/weather-suitable.schema.js";
+import { getWeatherAsync } from "./weather.service.js";
 
 const repository: IEquipmentRepository = new EquipmentRepositoryMemory();
 
@@ -18,7 +22,7 @@ export const addEquipment = async(equipmentInfo: CreateEquipmentRequest): Promis
 
   const existing = await repository.getBySerialNumber(equipmentInfo.serialNumber);
   if (existing) {
-    throw new Error("Сериалномер уже есть");
+    throw new Error("Сериал номер уже есть");
   }
   
   const equipment = Equipment.create(equipmentInfo);
@@ -31,17 +35,55 @@ export const getEquipments = async(request: getEquipmentsRequest): Promise<Equip
 };
 
 export const deleteEquipment = async(id: string): Promise<void> => {
+  const existing = await repository.getById(id);
+  if (!existing) {
+    throw new Error("Не найден");
+  }
+
   await repository.delete(id);
 };
 
 export const getEquipment = async(id: string): Promise<Equipment | undefined> => {
+  const existing = await repository.getById(id);
+  if (!existing) {
+    throw new Error("Не найден");
+  }
+
   return await repository.getById(id);
 };
 
 export const updateEquipment = async(id: string, equipmentInfo: UpdateEquipmentRequest): Promise<void> => {
-  let equipment = await getEquipment(id);
+  const equipment = await repository.getById(id);
+  if (!equipment) {
+    throw new Error("Не найден");
+  }
+
   if (equipment) {
     const updated = { ...equipment, ...equipmentInfo };
     await repository.update(updated);
   }
 };
+
+
+export const getEquipmentWeather = async(id: string): Promise<GetEquipmentWeatherResponse> => {
+  const response = new GetEquipmentWeatherResponse();
+  const equipment = await repository.getById(id);
+  if (!equipment) {
+    throw new Error("Не найден");
+  }
+
+  response.equipmentId = id;
+  response.location = equipment.location;
+
+  const daysWeather = await getWeatherAsync(equipment.location.lat, equipment.location.lon);
+  for (const weather of daysWeather) {
+    const parsed = weatherSuitableSchema.safeParse(weather);
+    console.log(parsed);
+    weather.suitable = parsed.success;
+    response.weather?.push(weather);
+  }
+
+  response.isWeatherWindowSuitable = response.weather.every(x => x.suitable);
+
+  return response;
+}
